requirement.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.5
colorama==0.4.4
cryptography==37.0.2
dateutils==0.6.12
Django==3.2.8
django-cors-headers==3.12.0
django-embed-video==1.4.4
django-rest-framework-braces==0.3.4
django-rest-knox==4.2.0
djangorestframework==3.13.1
ghasedakpack==0.1.9
idna==3.2
importlib-metadata==4.11.4
iniconfig==1.1.1
joblib==1.1.0
kavenegar==1.1.2
numpy==1.21.6
opencv-python==4.5.5.64
packaging==21.0
pandas==1.3.5
Pillow==9.1.1
pluggy==1.0.0
py==1.10.0
pycparser==2.21
PyJWT==2.4.0
pyotp==2.6.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.1
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
twilio==7.9.1
typing_extensions==4.1.1
urllib3==1.26.6
zipp==3.8.0
